backend/requirements.txt
Django==5.2.13
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
